feat: skeleton

+++ features/00_go/install.sh
@@ -1,7 +1,15 @@
 #!/bin/bash
 
-rm -rf /usr/local/go
+source "$(dirname "$0")/../utils.sh"
+
 GO_VERSION=1.26.5
+
+if check_command_version go "$GO_VERSION"; then
+    echo "==> go $GO_VERSION already installed, skipping"
+    exit 0
+fi
+
+rm -rf /usr/local/go
 
 TMP_FOLDER=$(mktemp -d)
